ai-pdf-summarizer/requirements.txt
streamlit
PyPDF2
reportlab
python-dotenv
groq
